Fix bug when seller releases funds before buyer sends fiat. Added translations to yaml.
