+++ locales/pt.yaml
@@ -35,6 +35,8 @@ invoice_payment_request: |
 
   Reputação do comprador: ${rate}
 
+  ⏱ Buyer usually sends fiat ${buyerResponse} minutes after accepting an offer
+
   Por favor, pague esta fatura para iniciar seu processo de venda, esta fatura expirará ${expirationTime} minutos
 pending_sell: |
   📝 Sua oferta foi publicada no ${channel} canal
@@ -112,6 +114,8 @@ someone_took_your_order: |
   Se por algum motivo o seu processador de pagamento coloca o pagamento em pausa e os fundos não chegam em menos de ${expirationTime}, Os SATs retornarão ao vendedor, colocando o comprador em risco e não poderei forçar o vendedor a enviar os sats novamente.
 
   Reputação do vendedor: ${rate}
+
+  ⏱ Seller usually releases sats within ${sellerResponse} minutes after sending fiat.
 
   Se você concorda com o exposto acima, pressione o botão para continuar 👇
 you_took_someone_order: |
@@ -467,3 +471,5 @@ bot_kicked: Bot was kicked from chat, must be member and admin
 chat_not_found: Chat not found, are you sure you added the bot to the group/channel?
 not_member: You are not member of that chat
 upgraded_to_supergroup: The chat was upgraded to 'supergroup' and the Id has changed, check the chat Id again
+buyer_time_response: Buyer responds in ${buyerResponse} min
+seller_time_response: Seller responds in ${sellerResponse} min
