+++ locales/ru.yaml
@@ -35,6 +35,8 @@ invoice_payment_request: |
 
   Buyer Reputation: ${rate}
 
+  ⏱ Buyer usually sends fiat ${buyerResponse} minutes after accepting an offer.
+
   Пожалуйста, оплатите этот счет, чтобы начать продажу. Срок действия этого счета истекает через ${expirationTime} минут
 pending_sell: |
   📝 Разместил заявку в канале ${channel}
@@ -111,6 +113,8 @@ someone_took_your_order: |
   Если по какой-либо причине Ваша платежная система задержит перевод, и средства не поступят в течение ${expirationTime}, сатоши будут автоматически возвращены продавцу. В этом случае покупатель несет риск, что мы не сможем убедить продавца выслать сатоши заново.
 
   Seller Reputation: ${rate}
+
+  ⏱ Seller usually releases sats within ${sellerResponse} minutes after sending fiat.
 
   Если Вы согласны с таким условием, нажмите кнопку Продолжить 👇
 you_took_someone_order: |
@@ -469,3 +473,5 @@ bot_kicked: Bot was kicked from chat, must be member and admin
 chat_not_found: Chat not found, are you sure you added the bot to the group/channel?
 not_member: You are not member of that chat
 upgraded_to_supergroup: The chat was upgraded to 'supergroup' and the Id has changed, check the chat Id again
+buyer_time_response: Buyer responds in ${buyerResponse} min
+seller_time_response: Seller responds in ${sellerResponse} min
